Create README.md
TerpBuy, a global business-to-consumer and business-to-business platform. It is based in and operates out of College Park, Maryland, USA, and has a distribution facility in Mumbai, India. The company is looking for insights on different aspects of its customers, products, departments, and orders. To help the company, I have worked on exporting data from its data warehouse and performed data analysis